feat: error logging
